+++ mobile/android/android-components/docs/changelog.md
@@ -21,10 +21,13 @@ permalink: /changelog/
 
 * **concept-engine**
   * Added `EngineSession.goToHistoryIndex` to jump to a specific index in a session's history.
-  
+
 * **service-location**
   * `LocationService.hasRegionCached()` is introduced to query if the region is already cached and a long running operation to fetch the region is not needed.
-  
+
+* **browser-state**
+  * Added map of `SessionState.contextId` and their respective `ContainerState` in `BrowserState` to track the state of a container.
+
 # 49.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v48.0.0...v49.0.0)